Large update

+++ MatlabFiles/Behavior/bit2num.m
@@ -4,5 +4,5 @@ A=0;
 S=size(B);
 
 for i=1:S(2)
-    A=A+B(S(2)-i+1)*(2^(i-1));
+    A=A+double(B(S(2)-i+1)*(2^(i-1)));
 end
